# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OSINT Research Tool API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SearchType = typeof SearchType[keyof typeof SearchType];


export const SearchType = {
  domain: 'domain',
  ip: 'ip',
  username: 'username',
  email: 'email',
  phone: 'phone',
} as const;

export type SearchStatus = typeof SearchStatus[keyof typeof SearchStatus];


export const SearchStatus = {
  pending: 'pending',
  completed: 'completed',
  failed: 'failed',
} as const;

/**
 * @nullable
 */
export type SearchResults = { [key: string]: unknown } | null;

export interface Search {
  id: number;
  query: string;
  type: SearchType;
  status: SearchStatus;
  /** @nullable */
  results?: SearchResults;
  createdAt: string;
}

export type SearchInputType = typeof SearchInputType[keyof typeof SearchInputType];


export const SearchInputType = {
  domain: 'domain',
  ip: 'ip',
  username: 'username',
  email: 'email',
  phone: 'phone',
} as const;

export interface SearchInput {
  /** @minLength 1 */
  query: string;
  type: SearchInputType;
}

export type SearchStatsByType = {[key: string]: number};

export interface SearchStats {
  total: number;
  byType: SearchStatsByType;
  recentCount: number;
}

export type TargetType = typeof TargetType[keyof typeof TargetType];


export const TargetType = {
  domain: 'domain',
  ip: 'ip',
  username: 'username',
  email: 'email',
  phone: 'phone',
  person: 'person',
} as const;

export interface Target {
  id: number;
  name: string;
  value: string;
  type: TargetType;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  tags?: string | null;
  createdAt: string;
}

export type TargetInputType = typeof TargetInputType[keyof typeof TargetInputType];


export const TargetInputType = {
  domain: 'domain',
  ip: 'ip',
  username: 'username',
  email: 'email',
  phone: 'phone',
  person: 'person',
} as const;

export interface TargetInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  value: string;
  type: TargetInputType;
  notes?: string;
  tags?: string;
}

export interface TargetUpdate {
  name?: string;
  notes?: string;
  tags?: string;
}

export interface PublicPost {
  id: string;
  platform: string;
  author: string;
  /** @nullable */
  authorHandle?: string | null;
  content: string;
  publishedAt: string;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  likes?: number | null;
  /** @nullable */
  shares?: number | null;
}

export type PostSearchInputPlatform = typeof PostSearchInputPlatform[keyof typeof PostSearchInputPlatform];


export const PostSearchInputPlatform = {
  all: 'all',
  twitter: 'twitter',
  reddit: 'reddit',
  mastodon: 'mastodon',
} as const;

export interface PostSearchInput {
  /** @minLength 1 */
  query: string;
  platform?: PostSearchInputPlatform;
  limit?: number;
}

export interface TrendingTopic {
  topic: string;
  count: number;
  /** @nullable */
  category?: string | null;
}

export type ListSearchesParams = {
type?: ListSearchesType;
limit?: number;
};

export type ListSearchesType = typeof ListSearchesType[keyof typeof ListSearchesType];


export const ListSearchesType = {
  domain: 'domain',
  ip: 'ip',
  username: 'username',
  email: 'email',
  phone: 'phone',
  all: 'all',
} as const;