# lib/api-zod/src/generated/types/postSearchInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OSINT Research Tool API
 * OpenAPI spec version: 0.1.0
 */
import type { PostSearchInputPlatform } from './postSearchInputPlatform';

export interface PostSearchInput {
  /** @minLength 1 */
  query: string;
  platform?: PostSearchInputPlatform;
  limit?: number;
}