# lib/api-zod/src/generated/types/publicPost.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OSINT Research Tool API
 * OpenAPI spec version: 0.1.0
 */

export interface PublicPost {
  id: string;
  platform: string;
  author: string;
  /** @nullable */
  authorHandle?: string | null;
  content: string;
  publishedAt: string;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  likes?: number | null;
  /** @nullable */
  shares?: number | null;
}