# lib/api-zod/src/generated/types/searchInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OSINT Research Tool API
 * OpenAPI spec version: 0.1.0
 */
import type { SearchInputType } from './searchInputType';

export interface SearchInput {
  /** @minLength 1 */
  query: string;
  type: SearchInputType;
}