# lib/api-zod/src/generated/types/searchStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OSINT Research Tool API
 * OpenAPI spec version: 0.1.0
 */
import type { SearchStatsByType } from './searchStatsByType';

export interface SearchStats {
  total: number;
  byType: SearchStatsByType;
  recentCount: number;
}