# lib/api-zod/src/generated/types/target.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OSINT Research Tool API
 * OpenAPI spec version: 0.1.0
 */
import type { TargetType } from './targetType';

export interface Target {
  id: number;
  name: string;
  value: string;
  type: TargetType;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  tags?: string | null;
  createdAt: string;
}