# lib/api-zod/src/generated/types/targetInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OSINT Research Tool API
 * OpenAPI spec version: 0.1.0
 */
import type { TargetInputType } from './targetInputType';

export interface TargetInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  value: string;
  type: TargetInputType;
  notes?: string;
  tags?: string;
}