# lib/api-zod/src/generated/types/trendingTopic.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OSINT Research Tool API
 * OpenAPI spec version: 0.1.0
 */

export interface TrendingTopic {
  topic: string;
  count: number;
  /** @nullable */
  category?: string | null;
}